All projects updated to v143

+++ src/materialsystem/stdshaders/water.cpp
@@ -506,9 +506,9 @@ BEGIN_VS_SHADER( SDK_Water_DX90,
 			float cheapWaterEndDistance = params[CHEAPWATERENDDISTANCE]->GetFloatValue();
 			float cheapWaterParams[4] = 
 			{
-				cheapWaterStartDistance * VSHADER_VECT_SCALE,
-				cheapWaterEndDistance * VSHADER_VECT_SCALE,
-				PSHADER_VECT_SCALE / ( cheapWaterEndDistance - cheapWaterStartDistance ),
+				cheapWaterStartDistance * (float)VSHADER_VECT_SCALE,
+				cheapWaterEndDistance * (float)VSHADER_VECT_SCALE,
+				(float)PSHADER_VECT_SCALE / ( cheapWaterEndDistance - cheapWaterStartDistance ),
 				cheapWaterStartDistance / ( cheapWaterEndDistance - cheapWaterStartDistance ),
 			};
 			pShaderAPI->SetPixelShaderConstant( 1, cheapWaterParams );
